requirements.txt
awsebcli==3.15.0
blessed==1.15.0
botocore==1.12.130
cached-property==1.5.1
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
coverage==4.5.2
Django==2.1.3
djangorestframework==3.9.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
gunicorn==19.9.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
OWSLib==0.17.0
pathspec==0.5.9
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyproj==1.9.5.1
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.1
wcwidth==0.1.7
websocket-client==0.56.0
